펫 추가&수정 때 강아지 몸무게 소숫점으로 입력받게 변경 결과값 반올림하게 변경

+++ app/src/main/java/com/example/porject/myListAdapter.kt
@@ -39,7 +39,7 @@ class myListAdapter (val context: Context, val myPetList: MutableList<myPetType>
                     if ((document["petType"] as Long) != getpetType) {
                         continue
                     }
-                    if ((document["petWeight"] as Long) != getpetWeight){
+                    if ((document["petWeight"] as Double) != getpetWeight){
                         continue
                     }
                     if ((document["userUID"] as String?) != getuserUid) {
@@ -67,7 +67,7 @@ class myListAdapter (val context: Context, val myPetList: MutableList<myPetType>
                     if ((document["petType"] as Long) != getpetType) {
                         continue
                     }
-                    if ((document["petWeight"] as Long) != getpetWeight){
+                    if ((document["petWeight"] as Double) != getpetWeight){
                         continue
                     }
                     if ((document["userUID"] as String?) != getuserUid) {
