+++ app/src/main/java/com/example/porject/myPetType.kt
@@ -1,3 +1,3 @@
 package com.example.porject
 
-data class myPetType(var petName: String = " ", var petType: Long = 0, var petWeight: Long = 0, var userUID: String? = " ")
+data class myPetType(var petName: String = " ", var petType: Long = 0, var petWeight: Double = 0.0, var userUID: String? = " ")
